GetYearBounds implementation, bug fixes

+++ ALS.Glance.DataAgents/Interfaces/IALSGlanceDA.cs
@@ -10,13 +10,13 @@ namespace ALS.Glance.DataAgents.Interfaces
     public interface IALSGlanceDA : IDataAgent
     {
         Task<IEnumerable<DimPatient>> GetPatientsAsync(WebApiCredentials credentials, CancellationToken ct);
-        Task<DimPatient> GetPatientAsync(WebApiCredentials credentials, int id,CancellationToken ct);
+        Task<DimPatient> GetPatientAsync(WebApiCredentials credentials, int id, CancellationToken ct);
 
         Task<ApplicationSettings> GetSettingsAsync(WebApiCredentials credentials, string userId, CancellationToken ct);
 
         Task<AgeBounds> GetAgeBoundsAsync(WebApiCredentials credentials, CancellationToken ct);
 
-        Task<YearBounds> GetYearBoundsAsync(WebApiCredentials credentials, long id, CancellationToken ct);
+        Task<YearBounds> GetYearBoundsAsync(WebApiCredentials credentials, int id, CancellationToken ct);
 
         Task<IEnumerable<DimMuscle>> GetMusclesAsync(WebApiCredentials credentials, CancellationToken ct);
     }
